Move node implementations into each node type
