+++ src/node_types/lazy.js
@@ -1,3 +1,4 @@
+import { register } from '../processor'
 const TYPE = 'lazy'
 
 export default function lazy(factory) {
@@ -8,3 +9,9 @@ export default function lazy(factory) {
 }
 
 lazy.TYPE = TYPE
+
+
+register(TYPE, (next, context, node, state, props, ...args) => {
+    const child = node.factory(...args)
+    return next(context, child, state, props)
+})
