+++ src/node_types/loader.js
@@ -1,3 +1,5 @@
+import { register } from '../processor'
+import { selectIsReady } from '../actions-reducer'
 const TYPE = 'loader'
 
 export default function loader(options = {}) {
@@ -23,3 +25,26 @@ export default function loader(options = {}) {
 }
 
 loader.TYPE = TYPE
+
+register(TYPE, (next, context, node, state, props, ...args) => {
+    const id = node.id(...args)
+    if (typeof id !== 'string') {
+        throw new Error('Loader failed to return an id')
+    }
+    let value
+
+    const isReady = node.lazy === true || selectIsReady(state, id)
+
+    // Always queue the action. The component can choose whether or not to
+    // call it.
+    context.queue(id, node.action, args)
+
+    if (isReady) {
+        value = node.selector(state, ...args)
+    }
+
+    return {
+        isReady,
+        value,
+    }
+})
